switch contact info ontainer to separate file
